wip rework constraint
